requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psycopg2-binary>=2.9.9
requests>=2.31.0
httpx>=0.25.0
curl_cffi>=0.7.0
anthropic>=0.40.0
